rooms/requirements.txt
aiofiles==25.1.0
fastapi==0.124.0
Jinja2==3.1.6
PyJWT==2.10.1
python-multipart==0.0.20
tortoise-orm==0.25.1
uvicorn==0.38.0